==> 5_checking_covariates_balance.Rmd <==
---
title: "Checking Covariates Balance"
description: |
  Comparing days with Wind Blowing from the North-East to Other Directions. Adjusting for calendar indicators and other weather variables.
author:
  - name: Léo Zabrocki 
    url: https://lzabrocki.github.io/
    affiliation: Paris School of Economics
    affiliation_url: https://www.parisschoolofeconomics.eu/fr/zabrocki-leo/
  - name: Anna Alari 
    url: https://scholar.google.com/citations?user=MiFY320AAAAJ&hl=fr
    affiliation: ISGlobal
    affiliation_url: https://www.isglobal.org/
  - name: Tarik Benmarhnia
    url: https://profiles.ucsd.edu/tarik.benmarhnia
    affiliation: UCSD & Scripps Institute
    affiliation_url: https://benmarhniaresearch.ucsd.edu/
date: "`r Sys.Date()`"
output: 
    distill::distill_article:
      toc: true
      toc_depth: 3
---

<style>
body {
text-align: justify}
</style>

```{r setup, include=FALSE}
# code chunk option
knitr::opts_chunk$set(
  echo = TRUE,
  message = FALSE,
  warning = FALSE,
  layout="l-body-outset",
  dev = "CairoPNG",
  dpi = 600
)
```

In this document, we take great care providing all steps and R codes required to check whether our matching procedure achieved balance. We compare days where:

* treated units are days where winds blow from the North-East in *t*.
* control units are day winds blow from other directions in *t*.

We adjust for calendar indicators and weather confouding factors.

**Should you have any questions, need help to reproduce the analysis or find coding errors, please do not hesitate to contact me at leo.zabrocki@psemail.eu**

# Required Packages

To reproduce exactly the `3_script_checking_balance_figures.html` document, we first need to have installed:

* the [R](https://www.r-project.org/) programming language 
* [RStudio](https://rstudio.com/), an integrated development environment for R, which will allow you to knit the `3_script_checking_balance_figures.Rmd` file and interact with the R code chunks
* the [R Markdown](https://rmarkdown.rstudio.com/) package
* and the [Distill](https://rstudio.github.io/distill/) package which provides the template for this document. 

Once everything is set up, we have to load the following packages:

```{r}
# load required packages
library(knitr) # for creating the R Markdown document
library(here) # for files paths organization
library(tidyverse) # for data manipulation and visualization
library(ggridges) # for ridge density plots
library(Cairo) # for printing customed police of graphs
library(patchwork) # combining plots
```

We finally load our custom `ggplot2` theme for graphs:

```{r}
# load ggplot custom theme
source(here::here(
  "inputs",
  "2.functions",
  "script_theme_tufte.R"
))
# define nice colors
my_blue <- "#0081a7"
my_orange <- "#fb8500"
```

The theme is based on the fantastic [hrbrthemes](https://hrbrmstr.github.io/hrbrthemes/index.html) package. If you do not want to use this theme or are unable to install it because of fonts issues, you can use the `theme_bw()` already included in the `ggplot2` package.

# Preparing the Data

We load the matched data:

```{r}
# load matched data
data_matched <-
  readRDS(here::here("inputs", "1.data", "5.matched_data", "matched_data.Rds"))
```

# Figures for Covariates Distribution for Treated and Control Units

We check whether coviariates balance was achieved with the thresholds we defined for our matching procedure. We plot distributions of weather and calendar variables (Lags 0-1) and pollutants (Lag 1) for treated and control groups.

### Weather Covariates

For continuous weather covariates, we draw boxplots for treated and control groups:

```{r, fig.width=10, fig.height=10, code_folding="Please show me the code!"}
# we select control variables and store them in a long dataframe
data_weather_continuous_variables <- data_matched %>%
  select(
    temperature_average,
    temperature_average_lag_1,
    humidity_average,
    humidity_average_lag_1,
    wind_speed,
    wind_speed_lag_1,
    is_treated
  ) %>%
  pivot_longer(
    cols = -c(is_treated),
    names_to = "variable",
    values_to = "values"
  ) %>%
  mutate(
    new_variable = NA %>%
      ifelse(
        str_detect(variable, "temperature_average"),
        "Average Temperature (°C)",
        .
      ) %>%
      ifelse(
        str_detect(variable, "humidity_average"),
        "Humidity Average (%)",
        .
      ) %>%
      ifelse(str_detect(variable, "wind_speed"), "Wind Speed (m/s)", .)
  ) %>%
  mutate(time = "in t" %>%
           ifelse(str_detect(variable, "lag_1"), "in t-1", .)) %>%
  mutate(variable = paste(new_variable, time, sep = " ")) %>%
  mutate(is_treated = if_else(is_treated == TRUE, "Treated", "Control"))

graph_boxplot_continuous_weather <-
  ggplot(data_weather_continuous_variables,
         aes(x = is_treated, y = values, colour = is_treated)) +
  geom_violin(size = 1) +
  geom_boxplot(width = 0.1, outlier.shape = NA) +
  scale_color_manual(values = c(my_blue, my_orange)) +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 10)) +
  ylab("Covariate Value") +
  xlab("") +
  labs(colour = "Units Status:") +
  facet_wrap( ~ variable, scale = "free", ncol = 2) +
  theme_tufte() +
  theme(
    legend.position = "top",
    legend.justification = "left",
    legend.direction = "horizontal"
  )

# we print the graph
graph_boxplot_continuous_weather

# save the graph
ggsave(
  graph_boxplot_continuous_weather,
  filename = here::here(
    "inputs", "3.outputs",
    "2.matching_analysis",
    "graph_boxplot_continuous_weather.pdf"
  ),
  width = 20,
  height = 20,
  units = "cm",
  device = cairo_pdf
)
````

For the rainfall duration and the wind direction categories, we plot the proportions:

```{r, fig.width=10, fig.height=5, code_folding="Please show me the code!"}
# we select the rainfall variables
data_weather_categorical <- data_matched %>%
  select(
    rainfall_duration,
    rainfall_duration_lag_1,
    wind_direction_categories_lag_1,
    is_treated
  ) %>%
  mutate_all( ~ as.character(.)) %>%
  pivot_longer(
    cols = -c(is_treated),
    names_to = "variable",
    values_to = "values"
  ) %>%
  # group by is_treated, variable and values
  group_by(is_treated, variable, values) %>%
  # compute the number of observations
  summarise(n = n()) %>%
  # compute the proportion
  mutate(freq = round(n / sum(n) * 100, 0)) %>%
  ungroup() %>%
  filter(!(
    variable %in% c("rainfall_duration", "rainfall_duration_lag_1") &
      values == "False"
  )) %>%
  mutate(
    new_variable = NA %>%
      ifelse(str_detect(variable, "wind"), "Wind Direction", .) %>%
      ifelse(str_detect(variable, "rainfall"), "Rainfall Duration", .)
  ) %>%
  mutate(time = "\nin t" %>%
           ifelse(str_detect(variable, "lag_1"), "\nin t-1", .)) %>%
  mutate(variable = paste(new_variable, time, sep = " ")) %>%
  mutate(is_treated = if_else(is_treated == TRUE, "Treated", "Control"))


# build the graph for wind direction
graph_categorical_wd_weather <- data_weather_categorical %>%
  filter(new_variable == "Wind Direction") %>%
  ggplot(., aes(x = freq, y = values, fill = is_treated)) +
  geom_point(shape = 21,
             size = 4,
             alpha = 0.8) +
  geom_vline(xintercept = 0, size = 0.3) +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 10)) +
  scale_fill_manual(values = c(my_blue, my_orange)) +
  facet_wrap( ~ variable, scales = "free") +
  xlab("Proportion (%)") +
  ylab("") +
  labs(fill = "Units Status:") +
  theme_tufte() +
  theme(
    legend.position = "top",
    legend.justification = "left",
    legend.direction = "horizontal"
  )

# we print the graph
graph_categorical_wd_weather

# build the graph for rainfall dummy
graph_categorical_rainfall_weather <- data_weather_categorical %>%
  filter(new_variable == "Rainfall Duration") %>%
  ggplot(., aes(x = freq, y = values, fill = is_treated)) +
  geom_point(shape = 21,
             size = 4,
             alpha = 0.8) +
  geom_vline(xintercept = 0, size = 0.3) +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 10)) +
  scale_fill_manual(values = c(my_blue, my_orange)) +
  facet_wrap( ~ variable) +
  xlab("Proportion (%)") +
  ylab("") +
  labs(fill = "Units Status:") +
  theme_tufte() +
  theme(
    legend.position = "top",
    legend.justification = "left",
    legend.direction = "horizontal"
  )

# we print the graph
graph_categorical_rainfall_weather

# combine plots
graph_categorical_weather <-
  graph_categorical_wd_weather / graph_categorical_rainfall_weather +
  plot_annotation(tag_levels = 'A') &
  theme(plot.tag = element_text(size = 20, face = "bold"))

# save the graph
ggsave(
  graph_categorical_weather,
  filename = here::here(
    "inputs", "3.outputs",
    "2.matching_analysis",
    "graph_categorical_weather.pdf"
  ),
  width = 20,
  height = 20,
  units = "cm",
  device = cairo_pdf
)
````

### Pollutants

For pollutants lag 1, we draw boxplots for treated and control groups:

```{r, fig.width=10, fig.height=5, code_folding="Please show me the code!"}
# we select control variables and store them in a long dataframe
data_pollutant_variables <- data_matched %>%
  select(mean_no2_lag_1:mean_pm25_lag_1,
         is_treated) %>%
  # transform the data to long to compute the proportion of observations for each variable
  pivot_longer(
    cols = -c(is_treated),
    names_to = "variable",
    values_to = "values"
  ) %>%
  mutate(is_treated = ifelse(is_treated == "TRUE", "Treated", "Control")) %>%
  mutate(
    pollutant = NA %>%
      ifelse(str_detect(variable, "no2"), "NO2", .) %>%
      ifelse(str_detect(variable, "o3"), "O3", .) %>%
      ifelse(str_detect(variable, "pm10"), "PM10", .) %>%
      ifelse(str_detect(variable, "pm25"), "PM2.5", .)
  ) %>%
  mutate(time = "in t-1") %>%
  mutate(variable = paste(pollutant, time, sep = " "))

# make graph
graph_boxplot_pollutants <- data_pollutant_variables %>%
  ggplot(., aes(x = is_treated, y = values, colour = is_treated)) +
  geom_violin(size = 0.5) +
  geom_boxplot(width = 0.1, outlier.shape = NA) +
  scale_color_manual(values = c(my_blue, my_orange)) +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 10)) +
  ylab("Concentration (µg/m³)") +
  xlab("") +
  labs(colour = "Units Status:") +
  facet_wrap( ~ variable, ncol = 4) +
  theme_tufte() +
  theme(
    legend.position = "top",
    legend.justification = "left",
    legend.direction = "horizontal"
  )

# we print the graph
graph_boxplot_pollutants

# save the graph
ggsave(
  graph_boxplot_pollutants,
  filename = here::here(
    "inputs", "3.outputs",
    "2.matching_analysis",
    "graph_boxplot_pollutants.pdf"
  ),
  width = 30,
  height = 10,
  units = "cm",
  device = cairo_pdf
)
````

### Calendar Indicator

For calendar variables such as the day of the week, bank days and holidays we matched strictly. We plot the proportions of observations belonging to each month by treatment status:

```{r, fig.width=10, fig.height=5, code_folding="Please show me the code!"}
# compute the proportions of observations belonging to each month by treatment status
data_month <- data_matched %>%
  mutate(is_treated = ifelse(is_treated == "TRUE", "Treated", "Control")) %>%
  select(month, is_treated) %>%
  mutate(
    month = recode(
      month,
      `1` = "January",
      `2` = "February",
      `3` = "March",
      `4` = "April",
      `5` = "May",
      `6` = "June",
      `7` = "July",
      `8` = "August",
      `9` = "September",
      `10` = "October",
      `11` = "November",
      `12` = "December"
    ) %>%
      fct_relevel(
        .,
        "January",
        "February",
        "March",
        "April",
        "May",
        "June",
        "July",
        "August",
        "September",
        "October",
        "November",
        "December"
      )
  ) %>%
  pivot_longer(.,-is_treated) %>%
  group_by(name, is_treated, value) %>%
  summarise(n = n()) %>%
  mutate(proportion = round(n / sum(n) * 100, 0)) %>%
  ungroup()

# we plot the data using cleveland dot plots
graph_month <-
  ggplot(data_month,
         aes(
           x = as.factor(value),
           y = proportion,
           colour = is_treated,
           group = is_treated
         )) +
  geom_line(size = 1) +
  scale_colour_manual(values = c(my_blue, my_orange),
                      guide = guide_legend(reverse = FALSE)) +
  ggtitle("Month") +
  ylab("Proportion (%)") +
  xlab("") +
  labs(colour = "Units Status:") +
  theme_tufte() +
  theme(
    legend.position = "top",
    legend.justification = "left",
    legend.direction = "horizontal"
  )

# we print the graph
graph_month
````

We plot the proportions of observations belonging to each year by treatment status:

```{r, fig.width=10, fig.height=5, code_folding="Please show me the code!"}
# compute the proportions of observations belonging to each year by treatment status
data_year <- data_matched %>%
  mutate(is_treated = ifelse(is_treated == "TRUE", "Treated", "Control")) %>%
  select(year, is_treated) %>%
  pivot_longer(.,-is_treated) %>%
  group_by(name, is_treated, value) %>%
  summarise(n = n()) %>%
  mutate(proportion = round(n / sum(n) * 100, 0)) %>%
  ungroup()

# we plot the data using cleveland dot plots
graph_year <-
  ggplot(data_year,
         aes(
           x = as.factor(value),
           y = proportion,
           colour = is_treated,
           group = is_treated
         )) +
  geom_line(size = 1) +
  scale_colour_manual(values = c(my_blue, my_orange),
                      guide = guide_legend(reverse = FALSE)) +
  ggtitle("Year") +
  ylab("Proportion (%)") +
  xlab("") +
  labs(colour = "Units Status:") +
  theme_tufte() +
  theme(
    legend.position = "top",
    legend.justification = "left",
    legend.direction = "horizontal"
  )

# we print the graph
graph_year
````

We combine and save the two previous plots:

```{r, fig.width=28, fig.height=18, code_folding="Please show me the code!"}
# combine plots
graph_month_year <- graph_month / graph_year +
  plot_annotation(tag_levels = 'A') &
  theme(plot.tag = element_text(size = 20, face = "bold"))

# save the plot
ggsave(
  graph_month_year,
  filename = here::here("inputs", "3.outputs", "2.matching_analysis", "graph_month_year.pdf"),
  width = 20,
  height = 17,
  units = "cm",
  device = cairo_pdf
)
````
